refactoring: extract function to parse request path
