+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/AppController.java
@@ -61,6 +61,7 @@ import org.hisp.dhis.user.UserDetails;
 import org.hisp.dhis.util.DateUtils;
 import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
 import org.hisp.dhis.webapi.service.ContextService;
+import org.hisp.dhis.webapi.service.HttpServletRequestPathParser;
 import org.hisp.dhis.webapi.utils.ContextUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.Resource;
@@ -113,7 +114,7 @@ public class AppController {
   public @ResponseBody Map<String, List<WebModule>> getWebModules(HttpServletRequest request) {
     checkForEmbeddedJettyRuntime(request);
 
-    String contextPath = ContextUtils.getContextPath(request);
+    String contextPath = HttpServletRequestPathParser.getContextPath(request);
     return Map.of("modules", getAccessibleAppMenu(contextPath));
   }
 
@@ -205,7 +206,7 @@ public class AppController {
   public void renderApp(
       @PathVariable("app") String app, HttpServletRequest request, HttpServletResponse response)
       throws IOException, WebMessageException {
-    String contextPath = ContextUtils.getContextPath(request);
+    String contextPath = HttpServletRequestPathParser.getContextPath(request);
     App application = appManager.getApp(app, contextPath);
 
     // Get page requested
