+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/SystemController.java
@@ -68,6 +68,7 @@ import org.hisp.dhis.user.CurrentUser;
 import org.hisp.dhis.user.CurrentUserUtil;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.webapi.mvc.annotation.ApiVersion;
+import org.hisp.dhis.webapi.service.HttpServletRequestPathParser;
 import org.hisp.dhis.webapi.utils.ContextUtils;
 import org.hisp.dhis.webapi.webdomain.CodeList;
 import org.hisp.dhis.webapi.webdomain.ObjectCount;
@@ -236,7 +237,7 @@ public class SystemController {
       @CurrentUser User currentUser) {
     SystemInfo info =
         systemService.getSystemInfo().toBuilder()
-            .contextPath(ContextUtils.getContextPath(request))
+            .contextPath(HttpServletRequestPathParser.getContextPath(request))
             .userAgent(request.getHeader(ContextUtils.HEADER_USER_AGENT))
             .build();
