+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/user/UserController.java
@@ -113,7 +113,7 @@ import org.hisp.dhis.user.UserSetting;
 import org.hisp.dhis.user.UserSettingKey;
 import org.hisp.dhis.user.Users;
 import org.hisp.dhis.webapi.controller.AbstractCrudController;
-import org.hisp.dhis.webapi.utils.ContextUtils;
+import org.hisp.dhis.webapi.service.HttpServletRequestPathParser;
 import org.hisp.dhis.webapi.webdomain.WebMetadata;
 import org.hisp.dhis.webapi.webdomain.WebOptions;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -395,7 +395,7 @@ public class UserController extends AbstractCrudController<User> {
 
     if (!userService.sendRestoreOrInviteMessage(
         user,
-        ContextUtils.getContextPath(request),
+        HttpServletRequestPathParser.getContextPath(request),
         userService.getRestoreOptions(user.getRestoreToken()))) {
       throw new WebMessageException(error("Failed to send invite message"));
     }
@@ -424,7 +424,9 @@ public class UserController extends AbstractCrudController<User> {
 
     userService.prepareUserForInvite(user);
     userService.sendRestoreOrInviteMessage(
-        user, ContextUtils.getContextPath(request), RestoreOptions.RECOVER_PASSWORD_OPTION);
+        user,
+        HttpServletRequestPathParser.getContextPath(request),
+        RestoreOptions.RECOVER_PASSWORD_OPTION);
   }
 
   @SuppressWarnings("unchecked")
@@ -794,7 +796,7 @@ public class UserController extends AbstractCrudController<User> {
         && importReport.getStats().getCreated() == 1
         && objectReport != null) {
       userService.sendRestoreOrInviteMessage(
-          user, ContextUtils.getContextPath(request), restoreOptions);
+          user, HttpServletRequestPathParser.getContextPath(request), restoreOptions);
 
       log.info(String.format("An invite email was successfully sent to: %s", user.getEmail()));
     }
