+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/service/DefaultContextService.java
@@ -33,8 +33,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import javax.servlet.http.HttpServletRequest;
 import org.apache.commons.collections4.CollectionUtils;
@@ -47,58 +45,23 @@ import org.springframework.web.context.request.ServletRequestAttributes;
  */
 @Service
 public class DefaultContextService implements ContextService {
-  private static Pattern API_VERSION = Pattern.compile("(/api/(\\d+)?/)");
 
   @Override
   public String getServletPath() {
-    return getContextPath() + getRequest().getServletPath();
+    HttpServletRequest request = getRequest();
+    return HttpServletRequestPathParser.getContextPath(request) + request.getServletPath();
   }
 
   @Override
   public String getContextPath() {
     HttpServletRequest request = getRequest();
-    StringBuilder builder = new StringBuilder();
-    String xForwardedProto = request.getHeader("X-Forwarded-Proto");
-    String xForwardedPort = request.getHeader("X-Forwarded-Port");
-
-    if (xForwardedProto != null
-        && (xForwardedProto.equalsIgnoreCase("http")
-            || xForwardedProto.equalsIgnoreCase("https"))) {
-      builder.append(xForwardedProto);
-    } else {
-      builder.append(request.getScheme());
-    }
-
-    builder.append("://").append(request.getServerName());
-
-    int port;
-
-    try {
-      port = Integer.parseInt(xForwardedPort);
-    } catch (NumberFormatException e) {
-      port = request.getServerPort();
-    }
-
-    if (port != 80 && port != 443) {
-      builder.append(":").append(port);
-    }
-
-    builder.append(request.getContextPath());
-
-    return builder.toString();
+    return HttpServletRequestPathParser.getContextPath(request);
   }
 
   @Override
   public String getApiPath() {
     HttpServletRequest request = getRequest();
-    Matcher matcher = API_VERSION.matcher(request.getRequestURI());
-    String version = "";
-
-    if (matcher.find()) {
-      version = "/" + matcher.group(2);
-    }
-
-    return getServletPath() + version;
+    return HttpServletRequestPathParser.getApiPath(request);
   }
 
   @Override
