+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/service/HttpServletRequestPathParser.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright (c) 2004-2024, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ * Neither the name of the HISP project nor the names of its contributors may
+ * be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.webapi.service;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import javax.servlet.http.HttpServletRequest;
+
+public class HttpServletRequestPathParser {
+  private static final Pattern API_VERSION = Pattern.compile("(/api/(\\d+)?/)");
+
+  private HttpServletRequestPathParser() {
+    throw new IllegalStateException("Utility class");
+  }
+
+  public static String getApiPath(HttpServletRequest request) {
+    Matcher matcher = API_VERSION.matcher(request.getRequestURI());
+    String version = "";
+
+    if (matcher.find()) {
+      version = "/" + matcher.group(2);
+    }
+
+    return getServletPath(request) + version;
+  }
+
+  public static String getServletPath(HttpServletRequest request) {
+    return getContextPath(request) + request.getServletPath();
+  }
+
+  public static String getContextPath(HttpServletRequest request) {
+    StringBuilder builder = new StringBuilder();
+    String xForwardedProto = request.getHeader("X-Forwarded-Proto");
+    String xForwardedPort = request.getHeader("X-Forwarded-Port");
+
+    if (xForwardedProto != null
+        && (xForwardedProto.equalsIgnoreCase("http")
+            || xForwardedProto.equalsIgnoreCase("https"))) {
+      builder.append(xForwardedProto);
+    } else {
+      builder.append(request.getScheme());
+    }
+
+    builder.append("://").append(request.getServerName());
+
+    int port;
+
+    try {
+      port = Integer.parseInt(xForwardedPort);
+    } catch (NumberFormatException e) {
+      port = request.getServerPort();
+    }
+
+    if (port != 80 && port != 443) {
+      builder.append(":").append(port);
+    }
+
+    builder.append(request.getContextPath());
+    return builder.toString();
+  }
+}
