+++ dhis-2/dhis-web/dhis-web-commons/src/main/java/org/hisp/dhis/webportal/menu/action/GetModulesAction.java
@@ -35,7 +35,7 @@ import org.apache.struts2.ServletActionContext;
 import org.hisp.dhis.user.CurrentUserUtil;
 import org.hisp.dhis.user.User;
 import org.hisp.dhis.user.UserService;
-import org.hisp.dhis.webapi.utils.ContextUtils;
+import org.hisp.dhis.webapi.service.HttpServletRequestPathParser;
 import org.hisp.dhis.webportal.module.Module;
 import org.hisp.dhis.webportal.module.ModuleManager;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -52,7 +52,8 @@ public class GetModulesAction implements Action {
 
   @Override
   public String execute() throws Exception {
-    String contextPath = ContextUtils.getContextPath(ServletActionContext.getRequest());
+    String contextPath =
+        HttpServletRequestPathParser.getContextPath(ServletActionContext.getRequest());
 
     modules = manager.getAccessibleMenuModulesAndApps(contextPath);
